+++ pkg/providers/iam/testdata/TestDataRole/valid.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+
+data "akamai_iam_role" "test" {
+  role_id = 12345
+}